+++ micropython-modules/microlite/xalloc.h
@@ -15,9 +15,9 @@
 #ifndef __XALLOC_H__
 #define __XALLOC_H__
 #include <stdint.h>
-void *xalloc(uint32_t size);
-void *xalloc_try_alloc(uint32_t size);
-void *xalloc0(uint32_t size);
+void *xalloc(size_t size);
+void *xalloc_try_alloc(size_t size);
+void *xalloc0(size_t size);
 void xfree(void *mem);
-void *xrealloc(void *mem, uint32_t size);
+void *xrealloc(void *mem, size_t size);
 #endif // __XALLOC_H__
